- move selector functions into context - create re-useable year and quarter components - re-structure reports dir
