--- tdrs-frontend/src/components/Reports/TanfSspReports.jsx
@@ -1,160 +1,0 @@
-import React from 'react'
-import { useDispatch } from 'react-redux'
-import classNames from 'classnames'
-import { clearFileList } from '../../actions/reports'
-import { quarters, constructYearOptions } from './utils'
-import { FiscalQuarterExplainer } from './Explainers'
-import SectionFileUploadForm from '../FileUploadForms/SectionFileUploadForm'
-import SectionSubmissionHistory from '../SubmissionHistory/SectionSubmissionHistory'
-import SegmentedControl from '../SegmentedControl'
-import { useReportsContext } from './ReportsContext'
-
-const TanfSspReports = ({ stt, isRegionalStaff }) => {
-  const dispatch = useDispatch()
-
-  const {
-    yearInputValue,
-    setYearInputValue,
-    quarterInputValue,
-    setQuarterInputValue,
-    fileTypeInputValue,
-    selectedSubmissionTab,
-    setSelectedSubmissionTab,
-    setLocalAlertState,
-    setReprocessedModalVisible,
-    setReprocessedDate,
-    headerRef,
-  } = useReportsContext()
-
-  const selectYear = ({ target: { value } }) => {
-    setYearInputValue(value)
-    setLocalAlertState({ active: false, type: null, message: null })
-    dispatch(clearFileList({ fileType: fileTypeInputValue }))
-  }
-
-  const selectQuarter = ({ target: { value } }) => {
-    setQuarterInputValue(value)
-    setLocalAlertState({ active: false, type: null, message: null })
-    dispatch(clearFileList({ fileType: fileTypeInputValue }))
-  }
-
-  return (
-    <>
-      <div className="grid-row grid-gap">
-        <div className="mobile:grid-container desktop:padding-0 desktop:grid-col-auto">
-          <div
-            className={classNames('usa-form-group maxw-mobile margin-top-4')}
-          >
-            <label
-              className="usa-label text-bold margin-top-4"
-              htmlFor="reportingYears"
-            >
-              Fiscal Year (October - September)*
-              <select
-                className={classNames('usa-select maxw-mobile')}
-                name="reportingYears"
-                id="reportingYears"
-                onChange={selectYear}
-                value={yearInputValue}
-                aria-describedby="years-error-alert"
-              >
-                <option value="" disabled hidden>
-                  - Select Fiscal Year -
-                </option>
-                {constructYearOptions(2021)}
-              </select>
-            </label>
-          </div>
-          <div
-            className={classNames('usa-form-group maxw-mobile margin-top-4')}
-          >
-            <label
-              className="usa-label text-bold margin-top-4"
-              htmlFor="quarter"
-            >
-              Fiscal Quarter*
-              <select
-                className={classNames('usa-select maxw-mobile')}
-                name="quarter"
-                id="quarter"
-                onChange={selectQuarter}
-                value={quarterInputValue}
-                aria-describedby="quarter-error-alert"
-              >
-                <option value="" disabled hidden>
-                  - Select Quarter -
-                </option>
-                {Object.entries(quarters).map(
-                  ([quarter, quarterDescription]) => (
-                    <option value={quarter} key={quarter}>
-                      {quarterDescription}
-                    </option>
-                  )
-                )}
-              </select>
-            </label>
-          </div>
-        </div>
-        <div className="mobile:grid-container desktop:padding-0 desktop:grid-col-fill">
-          <FiscalQuarterExplainer />
-        </div>
-      </div>
-
-      {yearInputValue && quarterInputValue && stt && (
-        <>
-          <hr />
-          <h2
-            ref={headerRef}
-            className="font-serif-xl margin-top-5 margin-bottom-0 text-normal"
-            tabIndex="-1"
-          >
-            {`${stt.name} - ${fileTypeInputValue.toUpperCase()} - Fiscal Year ${yearInputValue} - ${quarters[quarterInputValue]}`}
-          </h2>
-
-          {isRegionalStaff ? (
-            <h3 className="font-sans-lg margin-top-5 margin-bottom-2 text-bold">
-              Submission History
-            </h3>
-          ) : (
-            <SegmentedControl
-              buttons={[
-                {
-                  id: 1,
-                  label: 'Current Submission',
-                  onSelect: () => setSelectedSubmissionTab(1),
-                },
-                {
-                  id: 2,
-                  label: 'Submission History',
-                  onSelect: () => setSelectedSubmissionTab(2),
-                },
-              ]}
-              selected={selectedSubmissionTab}
-            />
-          )}
-
-          {!isRegionalStaff && selectedSubmissionTab === 1 && (
-            <SectionFileUploadForm stt={stt} />
-          )}
-
-          {(isRegionalStaff || selectedSubmissionTab === 2) && (
-            <SectionSubmissionHistory
-              filterValues={{
-                quarter: quarterInputValue,
-                year: yearInputValue,
-                stt: stt,
-                file_type: fileTypeInputValue,
-              }}
-              reprocessedState={{
-                setModalVisible: setReprocessedModalVisible,
-                setDate: setReprocessedDate,
-              }}
-            />
-          )}
-        </>
-      )}
-    </>
-  )
-}
-
-export default TanfSspReports
